pyproject.toml
[project]
name = "notes-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg[binary]>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.128.0",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
